@@ -1,7 +1,5 @@
 <script>
-	import '../app.css';
+	let { children } = $props();
 </script>
 
-<slot></slot>
-
-<style></style>
+{@render children()}
