@@ -1,8 +1,11 @@
 <script lang="ts">
 	import ContextMenu from '$lib/ContextMenu.svelte';
 	import { openContextMenu } from '$lib/functions.js';
-	import type { ContextMenuOptions } from '$lib/store.js';
-
 </script>
 
 <ContextMenu class="border border-neutral-500 bg-neutral-100 shadow-lg" />
+
+<button oncontextmenu={(e) => openContextMenu(e, [{
+	label: 'Hello friends',
+	action: () => console.log('Hello friends')
+}])}>Hello friends</button>
